+++ CHANGES.md
@@ -21,7 +21,7 @@
 
 ## [WIP] 0.7.2 / 2018-January-??
 
-- [TODO] Add utility fn `keypin.util/clojurize-subst` for symbol/keyword variable substitution in EDN data
+- Add utility fn `keypin.util/clojurize-subst` for symbol/keyword variable substitution in EDN data
   - Symbol variable example: $foo   (looked up as string `"foo"`)
   - Keyword variable example: :$bar (looked up as keyword `:bar`)